Refactor to new router, and minor changes
