+++ studio/__main_unit__.py
@@ -50,6 +50,7 @@ from studio.app.common.routers import (
     dataview,
     experiment,
     files,
+    frontend_errors,
     internal,
     logs,
     outputs,
@@ -205,6 +206,7 @@ app.include_router(auth.router)
 app.include_router(internal.router)  # Uses internal secret auth, not JWT
 app.include_router(experiment.router, dependencies=[Depends(get_current_user)])
 app.include_router(files.router, dependencies=[Depends(get_current_user)])
+app.include_router(frontend_errors.router, dependencies=[Depends(get_current_user)])
 app.include_router(logs.router, dependencies=[Depends(get_current_user)])
 app.include_router(
     outputs.router, dependencies=[Depends(get_current_user_with_dataview_outputs_check)]
