+++ studio/tests/app/common/routers/test_frontend_errors_api_contract.py
@@ -26,7 +26,7 @@ from fastapi.testclient import TestClient
 
 from studio.__main_unit__ import app
 from studio.app.common.core.auth.auth_dependencies import get_current_user
-from studio.app.common.routers.users_me import _frontend_error_timestamps
+from studio.app.common.routers.frontend_errors import _frontend_error_timestamps
 
 # ============================================================================
 # Frontend Contract Definitions
